requirements.txt
streamlit
tensorflow
torch
transformers
pandas
numpy
Pillow
matplotlib
scikit-learn
openpyxl
